datagen

+++ fuse-datagen/src/test/java/com/kayhut/fuse/generator/helper/TestUtil.java
@@ -36,9 +36,11 @@ public class TestUtil {
 
 
     public static void cleanDirectory(File dir) {
-        for (File file: dir.listFiles()) {
-            if (file.isDirectory()) cleanDirectory(file);
-            file.delete();
+        if (dir.exists()) {
+            for (File file : dir.listFiles()) {
+                if (file.isDirectory()) cleanDirectory(file);
+                file.delete();
+            }
         }
     }
 }
